Imported Upstream version 4.1.2-dfsg

+++ src/VBox/Main/src-server/USBControllerImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: USBControllerImpl.cpp 35638 2011-01-19 19:10:49Z vboxsync $ */
+/* $Id: USBControllerImpl.cpp 38441 2011-08-12 16:25:06Z vboxsync $ */
 /** @file
  * Implementation of IUSBController.
  */
@@ -533,7 +533,8 @@ STDMETHODIMP USBController::InsertDeviceFilter(ULONG aPosition,
     filter->mInList = true;
 
     /* notify the proxy (only when it makes sense) */
-    if (filter->getData().mActive && Global::IsOnline(adep.machineState()))
+    if (filter->getData().mActive && Global::IsOnline(adep.machineState())
+        && filter->getData().mRemote.isMatch (false))
     {
         USBProxyService *service = m->pHost->usbProxyService();
         ComAssertRet(service, E_FAIL);
@@ -604,7 +605,8 @@ STDMETHODIMP USBController::RemoveDeviceFilter(ULONG aPosition,
     filter.queryInterfaceTo(aFilter);
 
     /* notify the proxy (only when it makes sense) */
-    if (filter->getData().mActive && Global::IsOnline(adep.machineState()))
+    if (filter->getData().mActive && Global::IsOnline(adep.machineState())
+        && filter->getData().mRemote.isMatch (false))
     {
         USBProxyService *service = m->pHost->usbProxyService();
         ComAssertRet(service, E_FAIL);
@@ -767,7 +769,8 @@ void USBController::rollback()
             {
                 /* notify the proxy (only when it makes sense) */
                 if ((*it)->getData().mActive &&
-                    Global::IsOnline (adep.machineState()))
+                    Global::IsOnline (adep.machineState())
+                    && (*it)->getData().mRemote.isMatch (false))
                 {
                     USBDeviceFilter *filter = *it;
                     Assert(filter->getId() != NULL);
@@ -791,7 +794,8 @@ void USBController::rollback()
                     m->llDeviceFilters->end())
                 {
                     /* notify the proxy (only when necessary) */
-                    if ((*it)->getData().mActive)
+                    if ((*it)->getData().mActive
+                            && (*it)->getData().mRemote.isMatch (false))
                     {
                         USBDeviceFilter *flt = *it; /* resolve ambiguity */
                         Assert(flt->getId() == NULL);
@@ -1013,17 +1017,20 @@ HRESULT USBController::onDeviceFilterChange (USBDeviceFilter *aFilter,
 
         if (aActiveChanged)
         {
-            /* insert/remove the filter from the proxy */
-            if (aFilter->getData().mActive)
+            if (aFilter->getData().mRemote.isMatch (false))
             {
-                ComAssertRet(aFilter->getId() == NULL, E_FAIL);
-                aFilter->getId() = service->insertFilter(&aFilter->getData().mUSBFilter);
-            }
-            else
-            {
-                ComAssertRet(aFilter->getId() != NULL, E_FAIL);
-                service->removeFilter(aFilter->getId());
-                aFilter->getId() = NULL;
+                /* insert/remove the filter from the proxy */
+                if (aFilter->getData().mActive)
+                {
+                    ComAssertRet(aFilter->getId() == NULL, E_FAIL);
+                    aFilter->getId() = service->insertFilter(&aFilter->getData().mUSBFilter);
+                }
+                else
+                {
+                    ComAssertRet(aFilter->getId() != NULL, E_FAIL);
+                    service->removeFilter(aFilter->getId());
+                    aFilter->getId() = NULL;
+                }
             }
         }
         else
@@ -1033,7 +1040,10 @@ HRESULT USBController::onDeviceFilterChange (USBDeviceFilter *aFilter,
                 /* update the filter in the proxy */
                 ComAssertRet(aFilter->getId() != NULL, E_FAIL);
                 service->removeFilter(aFilter->getId());
-                aFilter->getId() = service->insertFilter(&aFilter->getData().mUSBFilter);
+                if (aFilter->getData().mRemote.isMatch (false))
+                {
+                    aFilter->getId() = service->insertFilter(&aFilter->getData().mUSBFilter);
+                }
             }
         }
     }
@@ -1209,7 +1219,9 @@ HRESULT USBController::notifyProxy (bool aInsertFilters)
         USBDeviceFilter *flt = *it; /* resolve ambiguity (for ComPtr below) */
 
         /* notify the proxy (only if the filter is active) */
-        if (flt->getData().mActive)
+        if (flt->getData().mActive
+                && flt->getData().mRemote.isMatch (false) /* and if the filter is NOT remote */
+                )
         {
             if (aInsertFilters)
             {
